requirements.txt
flask==3.1.2
flask-cors==6.0.1
gunicorn>=23.0.0
requests>=2.32.5
python-dotenv>=1.0.0
PyYAML>=6.0.1
dropbox
psycopg2-binary
openai
pymupdf
openpyxl
pandas
pytesseract
fitz
weasyprint
email_validator
flask-sqlalchemy
flask-login
